+++ packages/components/src/index.js
@@ -8,3 +8,4 @@ export { default as Stepper } from './stepper';
 export { default as FlowNavigation } from './flowNavigation';
 export { default as Popover } from './popover';
 export { default as Money } from './money';
+export { default as Modal } from './modal';