+++ packages/components/src/modal/Backdrop/Backdrop.js
@@ -1,0 +1,5 @@
+import React from 'react';
+
+const Backdrop = () => <div className="modal-backdrop fade" />;
+
+export default Backdrop;